Refactor satpy.dataset to separate modules
