+++ satpy/dataset/dataid.py
@@ -1,6 +1,6 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
-# Copyright (c) 2015-2019 Satpy developers
+# Copyright (c) 2015-2020 Satpy developers
 #
 # This file is part of satpy.
 #
@@ -21,197 +21,15 @@ import logging
 import numbers
 import warnings
 from collections import namedtuple
-from collections.abc import Collection
 from contextlib import suppress
 from copy import copy, deepcopy
-from datetime import datetime
 from enum import IntEnum, Enum
-from functools import reduce, partial
-from operator import is_, eq
 
 import numpy as np
 
 logger = logging.getLogger(__name__)
 
 
-class TooManyResults(KeyError):
-    """Special exception when one key maps to multiple items in the container."""
-
-    pass
-
-
-class MetadataObject(object):
-    """A general metadata object."""
-
-    def __init__(self, **attributes):
-        """Initialize the class with *attributes*."""
-        self.attrs = attributes
-
-    @property
-    def id(self):
-        """Return the DataID of the object."""
-        try:
-            return self.attrs['_satpy_id']
-        except KeyError:
-            id_keys = self.attrs.get('_satpy_id_keys', minimal_default_keys_config)
-            return DataID(id_keys, **self.attrs)
-
-
-def combine_metadata(*metadata_objects, average_times=True):
-    """Combine the metadata of two or more Datasets.
-
-    If the values corresponding to any keys are not equal or do not
-    exist in all provided dictionaries then they are not included in
-    the returned dictionary.  By default any keys with the word 'time'
-    in them and consisting of datetime objects will be averaged. This
-    is to handle cases where data were observed at almost the same time
-    but not exactly.  In the interest of time, lazy arrays are compared by
-    object identity rather than by their contents.
-
-    Args:
-        *metadata_objects: MetadataObject or dict objects to combine
-        average_times (bool): Average any keys with 'time' in the name
-
-    Returns:
-        dict: the combined metadata
-
-    """
-    info_dicts = _get_valid_dicts(metadata_objects)
-
-    if len(info_dicts) == 1:
-        return info_dicts[0].copy()
-
-    shared_keys = _shared_keys(info_dicts)
-
-    return _combine_shared_info(shared_keys, info_dicts, average_times)
-
-
-def _get_valid_dicts(metadata_objects):
-    """Get the valid dictionaries matching the metadata_objects."""
-    info_dicts = []
-    for metadata_object in metadata_objects:
-        if isinstance(metadata_object, dict):
-            metadata_dict = metadata_object
-        elif hasattr(metadata_object, "attrs"):
-            metadata_dict = metadata_object.attrs
-        else:
-            continue
-        info_dicts.append(metadata_dict)
-    return info_dicts
-
-
-def _shared_keys(info_dicts):
-    key_sets = (set(metadata_dict.keys()) for metadata_dict in info_dicts)
-    return reduce(set.intersection, key_sets)
-
-
-def _combine_shared_info(shared_keys, info_dicts, average_times):
-    shared_info = {}
-    for key in shared_keys:
-        values = [info[key] for info in info_dicts]
-        if 'time' in key and isinstance(values[0], datetime) and average_times:
-            shared_info[key] = average_datetimes(values)
-        elif _are_values_combinable(values):
-            shared_info[key] = values[0]
-    return shared_info
-
-
-def average_datetimes(datetime_list):
-    """Average a series of datetime objects.
-
-    .. note::
-
-        This function assumes all datetime objects are naive and in the same
-        time zone (UTC).
-
-    Args:
-        datetime_list (iterable): Datetime objects to average
-
-    Returns: Average datetime as a datetime object
-
-    """
-    total = [datetime.timestamp(dt) for dt in datetime_list]
-    return datetime.fromtimestamp(sum(total) / len(total))
-
-
-def _are_values_combinable(values):
-    """Check if the *values* can be combined."""
-    if _contain_arrays(values):
-        return _all_arrays_equal(values)
-    elif _contain_collections_of_arrays(values):
-        # in the real world, the `ancillary_variables` attribute may be
-        # List[xarray.DataArray], this means our values are now
-        # List[List[xarray.DataArray]].
-        # note that this list_of_arrays check is also true for any
-        # higher-dimensional ndarray, but we only use this check after we have
-        # checked any_arrays so this false positive should have no impact
-        return _all_list_of_arrays_equal(values)
-    return _all_values_equal(values)
-
-
-def _contain_arrays(values):
-    return any([_is_array(value) for value in values])
-
-
-def _is_array(val):
-    """Check if val is an array."""
-    return hasattr(val, "__array__") and not np.isscalar(val)
-
-
-nan_allclose = partial(np.allclose, equal_nan=True)
-
-
-def _all_arrays_equal(arrays):
-    """Check if the arrays are equal.
-
-    If the arrays are lazy, just check if they have the same identity.
-    """
-    if hasattr(arrays[0], 'compute'):
-        return _all_identical(arrays)
-    else:
-        return _pairwise_all(nan_allclose, arrays)
-
-
-def _pairwise_all(func, values):
-    for value in values[1:]:
-        if not func(values[0], value):
-            return False
-    return True
-
-
-def _all_identical(values):
-    """Check that the identities of all values are the same."""
-    return _pairwise_all(is_, values)
-
-
-def _contain_collections_of_arrays(values):
-    return any(
-        [_is_non_empty_collection(value) and
-         _is_all_arrays(value)
-         for value in values])
-
-
-def _is_non_empty_collection(value):
-    return isinstance(value, Collection) and len(value) > 0
-
-
-def _is_all_arrays(value):
-    return all([_is_array(sub_value) for sub_value in value])
-
-
-def _all_list_of_arrays_equal(array_lists):
-    """Check that the lists of arrays are equal."""
-    for array_list in zip(*array_lists):
-        if not _all_arrays_equal(array_list):
-            return False
-    return True
-
-
-def _all_values_equal(values):
-    try:
-        return _pairwise_all(nan_allclose, values)
-    except TypeError:
-        return _pairwise_all(eq, values)
 
 
 def get_keys_from_config(common_id_keys, config):
@@ -846,270 +664,3 @@ def create_filtered_query(dataset_key, filter_query):
                 ds_dict.setdefault(key, value)
 
     return DataQuery.from_dict(ds_dict)
-
-
-def dataset_walker(datasets):
-    """Walk through *datasets* and their ancillary data.
-
-    Yields datasets and their parent.
-    """
-    for dataset in datasets:
-        yield dataset, None
-        for anc_ds in dataset.attrs.get('ancillary_variables', []):
-            try:
-                anc_ds.attrs
-                yield anc_ds, dataset
-            except AttributeError:
-                continue
-
-
-def replace_anc(dataset, parent_dataset):
-    """Replace *dataset* the *parent_dataset*'s `ancillary_variables` field."""
-    if parent_dataset is None:
-        return
-    id_keys = parent_dataset.attrs.get('_satpy_id_keys', dataset.attrs.get('_satpy_id_keys'))
-    current_dataid = DataID(id_keys, **dataset.attrs)
-    for idx, ds in enumerate(parent_dataset.attrs['ancillary_variables']):
-        if current_dataid == DataID(id_keys, **ds.attrs):
-            parent_dataset.attrs['ancillary_variables'][idx] = dataset
-            return
-
-
-def get_best_dataset_key(key, choices):
-    """Choose the "best" `DataID` from `choices` based on `key`.
-
-    To see how the keys are sorted, refer to `:meth:satpy.datasets.DataQuery.sort_dataids`.
-
-    This function assumes `choices` has already been filtered to only
-    include datasets that match the provided `key`.
-
-    Args:
-        key (DataQuery): Query parameters to sort `choices` by.
-        choices (iterable): `DataID` objects to sort through to determine
-                            the best dataset.
-
-    Returns: List of best `DataID`s from `choices`. If there is more
-             than one element this function could not choose between the
-             available datasets.
-
-    """
-    sorted_choices, distances = key.sort_dataids(choices)
-    if len(sorted_choices) == 0 or distances[0] is np.inf:
-        return []
-    else:
-        return [choice for choice, distance in zip(sorted_choices, distances) if distance == distances[0]]
-
-
-def filter_keys_by_dataset_query(dquery, key_container):
-    """Filer provided key iterable by the provided `DataQuery`.
-
-    Note: The `modifiers` attribute of `did` should be `None` to allow for
-          **any** modifier in the results.
-
-    Args:
-        dquery (DataQuery): Query parameters to match in the `key_container`.
-        key_container (iterable): Set, list, tuple, or dict of `DataID`
-                                  keys.
-
-    Returns (list): List of keys matching the provided parameters in no
-                    specific order.
-
-    """
-    return dquery.filter_dataids(key_container)
-
-
-def get_key(key, key_container, num_results=1, best=True, query=None,
-            **kwargs):
-    """Get the fully-specified key best matching the provided key.
-
-    Only the best match is returned if `best` is `True` (default). See
-    `get_best_dataset_key` for more information on how this is determined.
-
-    `query` is provided as a convenience to filter by multiple parameters
-    at once without having to filter by multiple `key` inputs.
-
-    Args:
-        key (DataID): DataID of query parameters to use for
-                         searching. Any parameter that is `None`
-                         is considered a wild card and any match is
-                         accepted.
-        key_container (dict or set): Container of DataID objects that
-                                     uses hashing to quickly access items.
-        num_results (int): Number of results to return. Use `0` for all
-                           matching results. If `1` then the single matching
-                           key is returned instead of a list of length 1.
-                           (default: 1)
-        best (bool): Sort results to get "best" result first
-                     (default: True). See `get_best_dataset_key` for details.
-        query (DataQuery): filter for the key which can contain for example:
-            resolution (float, int, or list): Resolution of the dataset in
-                                            dataset units (typically
-                                            meters). This can also be a
-                                            list of these numbers.
-            calibration (str or list): Dataset calibration
-                                    (ex.'reflectance'). This can also be a
-                                    list of these strings.
-            polarization (str or list): Dataset polarization
-                                        (ex.'V'). This can also be a
-                                        list of these strings.
-            level (number or list): Dataset level (ex. 100). This can also be a
-                                    list of these numbers.
-            modifiers (list): Modifiers applied to the dataset. Unlike
-                            resolution and calibration this is the exact
-                            desired list of modifiers for one dataset, not
-                            a list of possible modifiers.
-
-
-    Returns (list or DataID): Matching key(s)
-
-    Raises: KeyError if no matching results or if more than one result is
-            found when `num_results` is `1`.
-
-    """
-    key = create_filtered_query(key, query)
-
-    res = filter_keys_by_dataset_query(key, key_container)
-    if not res:
-        raise KeyError("No dataset matching '{}' found".format(str(key)))
-
-    if best:
-        res = get_best_dataset_key(key, res)
-
-    if num_results == 1 and not res:
-        raise KeyError("No dataset matching '{}' found".format(str(key)))
-    elif num_results == 1 and len(res) != 1:
-        raise TooManyResults("No unique dataset matching {}".format(str(key)))
-    elif num_results == 1:
-        return res[0]
-    elif num_results == 0:
-        return res
-    else:
-        return res[:num_results]
-
-
-class DatasetDict(dict):
-    """Special dictionary object that can handle dict operations based on dataset name, wavelength, or DataID.
-
-    Note: Internal dictionary keys are `DataID` objects.
-
-    """
-
-    def keys(self, names=False, wavelengths=False):
-        """Give currently contained keys."""
-        # sort keys so things are a little more deterministic (.keys() is not)
-        keys = sorted(super(DatasetDict, self).keys())
-        if names:
-            return (k.get('name') for k in keys)
-        elif wavelengths:
-            return (k.get('wavelength') for k in keys)
-        else:
-            return keys
-
-    def get_key(self, match_key, num_results=1, best=True, **dfilter):
-        """Get multiple fully-specified keys that match the provided query.
-
-        Args:
-            key (DataID): DataID of query parameters to use for
-                          searching. Any parameter that is `None`
-                          is considered a wild card and any match is
-                          accepted. Can also be a string representing the
-                          dataset name or a number representing the dataset
-                          wavelength.
-            num_results (int): Number of results to return. If `0` return all,
-                               if `1` return only that element, otherwise
-                               return a list of matching keys.
-            **dfilter (dict): See `get_key` function for more information.
-
-        """
-        return get_key(match_key, self.keys(), num_results=num_results,
-                       best=best, **dfilter)
-
-    def getitem(self, item):
-        """Get Node when we know the *exact* DataID."""
-        return super(DatasetDict, self).__getitem__(item)
-
-    def __getitem__(self, item):
-        """Get item from container."""
-        try:
-            # short circuit - try to get the object without more work
-            return super(DatasetDict, self).__getitem__(item)
-        except KeyError:
-            key = self.get_key(item)
-            return super(DatasetDict, self).__getitem__(key)
-
-    def get(self, key, default=None):
-        """Get value with optional default."""
-        try:
-            key = self.get_key(key)
-        except KeyError:
-            return default
-        return super(DatasetDict, self).get(key, default)
-
-    def __setitem__(self, key, value):
-        """Support assigning 'Dataset' objects or dictionaries of metadata."""
-        value_dict = value
-        if hasattr(value, 'attrs'):
-            # xarray.DataArray objects
-            value_dict = value.attrs
-        # use value information to make a more complete DataID
-        if not isinstance(key, DataID):
-            if not isinstance(value_dict, dict):
-                raise ValueError("Key must be a DataID when value is not an xarray DataArray or dict")
-            old_key = key
-            try:
-                key = self.get_key(key)
-            except KeyError:
-                if isinstance(old_key, str):
-                    new_name = old_key
-                else:
-                    new_name = value_dict.get("name")
-                # this is a new key and it's not a full DataID tuple
-                if new_name is None and value_dict.get('wavelength') is None:
-                    raise ValueError("One of 'name' or 'wavelength' attrs "
-                                     "values should be set.")
-                try:
-                    id_keys = value_dict['_satpy_id'].id_keys
-                except KeyError:
-                    try:
-                        id_keys = value_dict['_satpy_id_keys']
-                    except KeyError:
-                        id_keys = minimal_default_keys_config
-                value_dict['name'] = new_name
-                key = DataID(id_keys, **value_dict)
-                if hasattr(value, 'attrs') and 'name' not in value.attrs:
-                    value.attrs['name'] = new_name
-
-        # update the 'value' with the information contained in the key
-        try:
-            new_info = key.to_dict()
-        except AttributeError:
-            new_info = key
-        if isinstance(value_dict, dict):
-            value_dict.update(new_info)
-
-        if hasattr(value, 'attrs'):
-            if isinstance(key, DataID):
-                value.attrs['_satpy_id'] = key
-
-        return super(DatasetDict, self).__setitem__(key, value)
-
-    def contains(self, item):
-        """Check contains when we know the *exact* DataID."""
-        return super(DatasetDict, self).__contains__(item)
-
-    def __contains__(self, item):
-        """Check if item exists in container."""
-        try:
-            key = self.get_key(item)
-        except KeyError:
-            return False
-        return super(DatasetDict, self).__contains__(key)
-
-    def __delitem__(self, key):
-        """Delete item from container."""
-        try:
-            # short circuit - try to get the object without more work
-            return super(DatasetDict, self).__delitem__(key)
-        except KeyError:
-            key = self.get_key(key)
-            return super(DatasetDict, self).__delitem__(key)
