+++ satpy/readers/yaml_reader.py
@@ -40,8 +40,8 @@ from pyresample.geometry import StackedAreaDefinition, SwathDefinition
 from pyresample.boundary import AreaDefBoundary, Boundary
 from satpy.resample import get_area_def
 from satpy.config import recursive_dict_update
-from satpy.dataset import DataQuery, get_keys_from_config, default_id_keys_config, default_co_keys_config, DataID, \
-    get_key
+from satpy.dataset import DataQuery, DataID, get_key
+from satpy.dataset.dataid import get_keys_from_config, default_id_keys_config, default_co_keys_config
 from satpy import DatasetDict
 from satpy.resample import add_crs_xy_coords
 from trollsift.parser import globify, parse
