+++ satpy/tests/reader_tests/test_mersi2_l1b.py
@@ -258,7 +258,7 @@ class TestMERSI2L1BReader(unittest.TestCase):
         """Test loading data when all resolutions are available."""
         from satpy.tests.utils import make_dataid
         from satpy.readers import load_reader
-        from satpy.dataset import get_key
+        from satpy.dataset.data_dict import get_key
         filenames = [
             'tf2019071182739.FY3D-X_MERSI_0250M_L1B.HDF',
             'tf2019071182739.FY3D-X_MERSI_1000M_L1B.HDF',
@@ -415,7 +415,7 @@ class TestMERSI2L1BReader(unittest.TestCase):
         """Test loading data when only 1km resolutions are available."""
         from satpy.tests.utils import make_dataid
         from satpy.readers import load_reader
-        from satpy.dataset import get_key
+        from satpy.dataset.data_dict import get_key
         filenames = [
             'tf2019071182739.FY3D-X_MERSI_1000M_L1B.HDF',
             'tf2019071182739.FY3D-X_MERSI_GEO1K_L1B.HDF',
@@ -477,7 +477,7 @@ class TestMERSI2L1BReader(unittest.TestCase):
         """Test loading data when only 250m resolutions are available."""
         from satpy.tests.utils import make_dataid
         from satpy.readers import load_reader
-        from satpy.dataset import get_key
+        from satpy.dataset.data_dict import get_key
         filenames = [
             'tf2019071182739.FY3D-X_MERSI_0250M_L1B.HDF',
             'tf2019071182739.FY3D-X_MERSI_GEOQK_L1B.HDF',
